refactor: rearrange folders and files to scale better
